Refactor join page (#542)
* Add @storybook/addon-actions back to package.json

The package is installed with @storybook/addon-essentials, so it
doesn't affect package-lock.json. However, eslint warns that it
should be listed as dependency.

* Intercept dispatch in redux store decorator

In order to develop components in isolation, all async calls
should be intercepted and mocked.

* Rename Join page file

* Update relative paths

* Remove history parameter from JoinPage

History is no longer dependency of useEffects(), and it is passed
to the HeaderSection, but never used.

* Pass queueName from url params to HeaderSection

Header section is visible while loading spinner. If queue exists,
the name will be same as in url params.

* Add JoinPage stories

* Update QueueStats stories to storybook 6.x style

* Reuse sotries data in JoinPage stories

* Rename Join scss file

* Rename Join form

* Destrurcture props in JoinForm

* Fix react wornings in  JoinForm

Callback chained to the promise was attempting to change the form
after redirect, which was causing errors in the cosnole.

* Add joining case to JoinPage stories

This sotry desribes state after the user click Join button,
before the request is fulfilled.